# Removing the unsafe code from the stable hash algorithm.