--- a/src/Microsoft.VisualStudio.ProjectSystem.Managed/ProjectSystem/UpToDate/BuildUpToDateCheck.ItemHashing.cs
+++ b/src/Microsoft.VisualStudio.ProjectSystem.Managed/ProjectSystem/UpToDate/BuildUpToDateCheck.ItemHashing.cs
@@ -46,28 +46,27 @@ namespace Microsoft.VisualStudio.ProjectSystem.UpToDate
         /// </remarks>
         internal static int GetStableHashCode(string str)
         {
-            unsafe
+            int hash1 = 5381;
+            int hash2 = hash1;
+
+            int i = 0;
+            while (i < str.Length)
             {
-                fixed (char* src = str)
-                {
-                    int hash1 = 5381;
-                    int hash2 = hash1;
+                char c = str[i];
 
-                    int c;
-                    char* s = src;
-                    while ((c = s[0]) != 0)
-                    {
-                        hash1 = ((hash1 << 5) + hash1) ^ c;
-                        c = s[1];
-                        if (c == 0)
-                            break;
-                        hash2 = ((hash2 << 5) + hash2) ^ c;
-                        s += 2;
-                    }
+                hash1 = ((hash1 << 5) + hash1) ^ c;
 
-                    return hash1 + (hash2 * 1566083941);
-                }
+                i++;
+                if (i == str.Length)
+                    break;
+
+                c = str[i];
+                hash2 = ((hash2 << 5) + hash2) ^ c;
+
+                i++;
             }
+
+            return hash1 + (hash2 * 1566083941);
         }
     }
 }