@davkean 2022-08-02
System.String is UTF16 on every OS and this code doesn't run on Mac/Linux, so I don't follow the Mac/Linux comment?